requirements.txt
langchain
langchain-community
langchain-ollama
ollama
python-dotenv
fastapi
uvicorn